requirements.txt
python-dotenv >= 0.15.0
numpy >= 1.19.2
pandas >= 1.2.3
matplotlib~=3.3.4
nltk >= 3.4.0
praw >= 7.2.0
setuptools >= 52.0.0
joblib > 1.0.0
gensim >= 4.0.0
torch >= 1.8.1
tokenizers >= 0.10.1
transformers >= 4.4.2
tensorflow >= 2.0
keras
keras-bert
